Move: files for external services
* These files are rather 'services' than 'helpers'.
* The filenames have been changed to reflect the above point.
* The files have also been put into their own folder,
rather than being stored in the 'helpers' folder.
