+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bettertaskfilter",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "Creating a filter in Todoist, which creates a view where tasks are shown based on their priority and project.",
   "main": "dist/index.js",
   "scripts": {
