+++ src/helpers/TaskHelper.ts
@@ -1,5 +1,5 @@
-import * as TodoistProvider from './TodoistProvider';
+import * as TodoistService from '../services/TodoistService';
 
 export async function getTasks(bucketFilter: FilterInterface): Promise<TaskInterface[]> {
-    return await TodoistProvider.retrieveTasks(bucketFilter);
+    return await TodoistService.retrieveTasks(bucketFilter);
 }
